move mapping function to common
